rename: api type 디렉토리 추가
